refactor(View): change ContentView to QuestionView
